requirements.txt
babel==2.17.0
backrefs==5.8
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
ghp-import==2.1.0
idna==3.10
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-gen-files==0.5.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.8
Pygments==2.19.1
pymdown-extensions==10.15
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.3
six==1.17.0
urllib3==2.4.0
watchdog==6.0.0
